--- .expo-shared/README.md
@@ -1,11 +1,0 @@
-> Why do I have a folder named ".expo-shared" in my project?
-
-The ".expo-shared" folder is created when running commands that produce state that is intended to be shared with all developers on the project. For example, "npx expo-optimize".
-
-> What does the "assets.json" file contain?
-
-The "assets.json" file describes the assets that have been optimized through "expo-optimize" and do not need to be processed again.
-
-> Should I commit the ".expo-shared" folder?
-
-Yes, you should share the ".expo-shared" folder with your collaborators.
