+++ README.md
@@ -1,0 +1,16 @@
+# culectuber-app
+repositorio para la app del colectivero
+-	Crear Proyecto
+ npm install -g react-native-cli
+
+
+-	Instalar location
+ expo install expo-location
+
+
+-		Para usar Axios 
+1) npm install axios   "O"   2) yarn add axios
+
+
+-	instalar swal
+  npm i sweettalert
