+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "ColectuberApp",
-    "slug": "ColectuberApp",
+    "name": "login-basico",
+    "slug": "snack-595a3aa5-9cc1-4a05-b504-e7c955d584b6",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
@@ -17,26 +17,18 @@
       "**/*"
     ],
     "ios": {
-      "supportsTablet": true,
-      "infoPlist": {
-        "UIBackgroundModes": [
-          "location",
-          "fetch"
-        ]
-      }
+      "supportsTablet": true
     },
     "android": {
       "adaptiveIcon": {
         "foregroundImage": "./assets/adaptive-icon.png",
         "backgroundColor": "#FFFFFF"
       },
-      "permissions": [
-        "ACCESS_BACKGROUND_LOCATION"
-      ]
+      "package": "com.josesosa96.snack595a3aa59cc14a05b504e7c955d584b6"
     },
     "web": {
       "favicon": "./assets/favicon.png"
     },
-    "description": ""
+    "description": "\nModelo de login básico"
   }
 }
