+++ babel.config.js
@@ -2,10 +2,5 @@ module.exports = function(api) {
   api.cache(true);
   return {
     presets: ['babel-preset-expo'],
-    plugins: [
-      ['module-resolver',{
-          root: ['.']
-      }],
-    ],
   };
 };
