+++ components/BordeSuperior.js
@@ -1,12 +1,11 @@
-import {widthPercentageToDP as wp, heightPercentageToDP as hp} from 'react-native-responsive-screen';
 import React, { Component } from 'react';
 import {View, StyleSheet, Image } from 'react-native';
-//Borde
+
 export default class BordeSuperior extends Component {
   render() {
     return (
       <View style={styles.container}>
-        <Image style={styles.logo} source={require("src/components/login/icons/bordesuperior.png")}/>
+        <Image style={styles.logo} source={require("../assets/bordesuperior.png")}/>
       </View>
     );
   }
@@ -14,15 +13,20 @@ export default class BordeSuperior extends Component {
 
 const styles = StyleSheet.create({
   container: {
-    flex: 1,
     alignItems: 'center',
     justifyContent: 'center',
   },
+  paragraph: {
+    margin: 24,
+    marginTop: 0,
+    fontSize: 14,
+    fontWeight: 'bold',
+    textAlign: 'center',
+    color: '#34495e',
+  },
   logo: {
     backgroundColor: "#ffffff",
-    height: hp('10%'), 
-    width: wp('25%')   
-    // height:70, 
-    //width:320   
+    height: 70,
+    width: 320,
   }
 });
