+++ package.json
@@ -1,34 +1,26 @@
 {
-  "name": "colectuberapp",
-  "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "eject": "expo eject"
+    "eject": "expo eject",
+    "build": "expo build:android"
   },
   "dependencies": {
-    "@react-navigation/native": "^6.0.10",
-    "@react-navigation/native-stack": "^6.6.1",
-    "axios": "^0.26.1",
-    "expo": "~44.0.0",
-    "expo-location": "~14.0.1",
-    "expo-secure-store": "~11.1.0",
-    "expo-status-bar": "~1.2.0",
-    "expo-task-manager": "~10.1.0",
+    "@expo/vector-icons": "^12.0.0",
+    "expo": "~43.0.2",
+    "expo-status-bar": "~1.1.0",
+    "expo-updates": "~0.10.15",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
-    "react-native-responsive-screen": "^1.4.2",
-    "react-native-safe-area-context": "3.3.2",
-    "react-native-screens": "~3.10.1",
+    "react-native-elements": "0.18.5",
     "react-native-web": "0.17.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "babel-plugin-module-resolver": "^4.1.0"
+    "@babel/core": "^7.12.9"
   },
   "private": true
 }
