--- src/api/api.js
@@ -1,7 +1,0 @@
-import axios from "axios";
-
-const API = axios.create({
-    baseURL:"https://colectuber-backend.azurewebsites.net"
-});
-
-export default API;
