--- src/components/loading/loading.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { View, Text } from 'react-native';
-
-const Loading = ()=>{
-    return <View>
-        <Text>
-            Loading...
-        </Text>
-    </View>
-}
-
-export default Loading;
